Oops, forgot to re-initialize test objects before each test!
For some reason, the tests were still passing. Maybe it's dumb luck.
Maybe JUnit works in a way that I wasn't expecting.

Either way, that's too many "oops"es in a short span of time. That means
that my brain is starting to need a break. So...time to take a break.